add customisable delay for the didSave event
Most users do not edit the pier directly, preferring to edit in a
git repo and automatically copy the files over. However, this
breaks autocommit, as the files may not have been copied over yet.
This commit adds a delay so that the language server will wait a
fixed amount of time before informing the language-server gall app
of the didSave event. This is so that the user can ensure the files
have been copied.

See:
https://groups.google.com/a/urbit.org/forum/#!topic/dev/uUThpGgZHPQ